+++ pytket/docs/changelog.rst
@@ -9,13 +9,14 @@ Features:
 * Add ``OpType.CnRx`` and ``OpType.CnRz``.
 * Add ``AutoRebase`` and ``AutoSquash`` passes.
   Deprecate ``auto_rebase_pass`` and ``auto_squash_pass``.
+* Add new parameter to `remove_blank_wires` to allow to keep empty classical bits.
 * Improve depth of circuit produced by ``MultiplexedTensoredU2Box``.
 
 Fixes:
 
 * Allow barriers when dagger or transpose a circuit.
+* Keep blank classical wires when running `FlattenRelabelRegistersPass`
 * Handle Clifford-angle ``NPhasedX`` gates in Clifford resynthesis.
-
 
 1.28.0 (May 2024)
 -----------------