+++ pytket/docs/changelog.rst
@@ -10,6 +10,8 @@ Features:
 * Add ``AutoRebase`` and ``AutoSquash`` passes.
   Deprecate ``auto_rebase_pass`` and ``auto_squash_pass``.
 * Add new parameter to `remove_blank_wires` to allow to keep empty classical bits
+* Support classical transforms and predicates, and QASM registers, with up to 64
+  bits.
 
 Fixes: